+++ app/src/routes/projects/ProjectsTab/index.tsx
@@ -1,18 +1,8 @@
-import { useState } from 'react';
-import EmblaCarousel from './Carousel';
 import CreateProject from './Create';
-import { ProjectInfo } from './Info';
-
-export interface Project {
-    id: number;
-    img: string;
-    title: string;
-}
+import SelectProject from './Select';
+import { Project } from '/common/models/project';
 
 export function ProjectsTab() {
-    const [currentProjectIndex, setCurrentProjectIndex] = useState(0);
-    const [direction, setDirection] = useState(0);
-
     const PROJECTS: Project[] = [
         // { id: 0, img: 'https://picsum.photos/id/237/200/300', title: 'Airbnb.com' },
         // { id: 1, img: 'https://picsum.photos/id/238/300/200', title: 'Netflix Clone' },
@@ -22,31 +12,10 @@ export function ProjectsTab() {
         // { id: 5, img: 'https://picsum.photos/id/242/1000/1000', title: 'YC' },
     ];
 
-    const handleProjectChange = (index: number) => {
-        if (currentProjectIndex === index) {
-            return;
-        }
-        setDirection(index > currentProjectIndex ? 1 : -1);
-        setCurrentProjectIndex(index);
-    };
-
     return (
         <div className="flex h-[calc(100vh-5.5rem)] w-full">
             {PROJECTS.length === 0 && <CreateProject />}
-            {PROJECTS.length > 0 && (
-                <>
-                    <div className="w-3/5">
-                        <EmblaCarousel slides={PROJECTS} onSlideChange={handleProjectChange} />
-                    </div>
-                    <div className="w-2/5 flex flex-col justify-center items-start p-4 gap-6">
-                        <ProjectInfo
-                            projects={PROJECTS}
-                            currentProjectIndex={currentProjectIndex}
-                            direction={direction}
-                        />
-                    </div>
-                </>
-            )}
+            {PROJECTS.length > 0 && <SelectProject projects={PROJECTS} />}
         </div>
     );
 }
